chore: update agevents.png image file ```
Since this is a binary file change with no functional impact on the codebase, I've used the "chore" type. The description simply states what was changed (the image file was updated) without going into details about why, as this information isn't available in the diff and the change appears to be a simple asset update.